fix: optimize performance, battle.net token refresh

+++ commands/defaultAction.js
@@ -3,11 +3,18 @@ const childs = require("./childs")
 const mongo = require("../db");
 const getMostMatchingCard = require("../tools/getMostMatchingCard");
 
+function base64_decode(base64Image, file) {
+  fs.writeFileSync(file,base64Image);
+   console.log('******** File created from base64 encoded string ********');
+}
+
+
 async function defaultAction(message, args, blizzardToken, class_){
   let infoMessage = await message.channel.send("🔍 검색 중입니다...")
   let userConfig = await mongo.userModel.findOne({name:`${message.author.username}#${message.author.discriminator}`}).exec();
   let gamemode = userConfig ? userConfig.gamemode : "wild";
   const resCard = await getMostMatchingCard(message, args, gamemode, blizzardToken);
+  if (!resCard) return;
   
   await message.channel.send({files: [resCard.image]});
   infoMessage.delete();
@@ -27,7 +34,7 @@ async function defaultAction(message, args, blizzardToken, class_){
 }
 
 module.exports = {
-    name : 'defaultAction',
-    description : 'defaultAction',
-    execute : defaultAction
+  name : 'defaultAction',
+  description : 'defaultAction',
+  execute : defaultAction
 }
